+++ cli/src/commands/file/chmod.rs
@@ -48,6 +48,23 @@ pub(crate) struct ChmodArgs {
 }
 
 #[instrument(skip_all)]
+pub(crate) fn deprecated_cmd_chmod(
+    ui: &mut Ui,
+    command: &CommandHelper,
+    args: &ChmodArgs,
+) -> Result<(), CommandError> {
+    writeln!(
+        ui.warning_default(),
+        "`jj chmod` is deprecated; use `jj file chmod` instead, which is equivalent"
+    )?;
+    writeln!(
+        ui.warning_default(),
+        "`jj chmod` will be removed in a future version, and this will be a hard error"
+    )?;
+    cmd_chmod(ui, command, args)
+}
+
+#[instrument(skip_all)]
 pub(crate) fn cmd_chmod(
     ui: &mut Ui,
     command: &CommandHelper,