+++ cli/tests/test_file_print_command.rs
@@ -15,7 +15,7 @@
 use crate::common::TestEnvironment;
 
 #[test]
-fn test_cat() {
+fn test_print() {
     let test_env = TestEnvironment::default();
     test_env.jj_cmd_ok(test_env.env_root(), &["git", "init", "repo"]);
     let repo_path = test_env.env_root().join("repo");
@@ -27,19 +27,13 @@ fn test_cat() {
     std::fs::write(repo_path.join("dir").join("file2"), "c\n").unwrap();
 
     // Can print the contents of a file in a commit
-    let stdout = test_env.jj_cmd_success(&repo_path, &["cat", "file1", "-r", "@-"]);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["file", "print", "file1", "-r", "@-"]);
     insta::assert_snapshot!(stdout, @r###"
     a
     "###);
 
     // Defaults to printing the working-copy version
-    let stdout = test_env.jj_cmd_success(&repo_path, &["cat", "file1"]);
-    insta::assert_snapshot!(stdout, @r###"
-    b
-    "###);
-
-    // `print` is an alias for `cat`
-    let stdout = test_env.jj_cmd_success(&repo_path, &["print", "file1"]);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["file", "print", "file1"]);
     insta::assert_snapshot!(stdout, @r###"
     b
     "###);
@@ -50,32 +44,33 @@ fn test_cat() {
     } else {
         "dir\\file2"
     };
-    let stdout = test_env.jj_cmd_success(&repo_path, &["cat", subdir_file]);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["file", "print", subdir_file]);
     insta::assert_snapshot!(stdout, @r###"
     c
     "###);
 
     // Error if the path doesn't exist
-    let stderr = test_env.jj_cmd_failure(&repo_path, &["cat", "nonexistent"]);
+    let stderr = test_env.jj_cmd_failure(&repo_path, &["file", "print", "nonexistent"]);
     insta::assert_snapshot!(stderr, @r###"
     Error: No such path: nonexistent
     "###);
 
     // Can print files under the specified directory
-    let stdout = test_env.jj_cmd_success(&repo_path, &["cat", "dir"]);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["file", "print", "dir"]);
     insta::assert_snapshot!(stdout, @r###"
     c
     "###);
 
     // Can print multiple files
-    let stdout = test_env.jj_cmd_success(&repo_path, &["cat", "."]);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["file", "print", "."]);
     insta::assert_snapshot!(stdout, @r###"
     c
     b
     "###);
 
     // Unmatched paths should generate warnings
-    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["cat", "file1", "non-existent"]);
+    let (stdout, stderr) =
+        test_env.jj_cmd_ok(&repo_path, &["file", "print", "file1", "non-existent"]);
     insta::assert_snapshot!(stdout, @r###"
     b
     "###);
@@ -87,7 +82,7 @@ fn test_cat() {
     test_env.jj_cmd_ok(&repo_path, &["new"]);
     std::fs::write(repo_path.join("file1"), "c\n").unwrap();
     test_env.jj_cmd_ok(&repo_path, &["rebase", "-r", "@", "-d", "@--"]);
-    let stdout = test_env.jj_cmd_success(&repo_path, &["cat", "file1"]);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["file", "print", "file1"]);
     insta::assert_snapshot!(stdout, @r###"
     <<<<<<< Conflict 1 of 1
     %%%%%%% Changes from base to side #1
@@ -101,7 +96,7 @@ fn test_cat() {
 
 #[cfg(unix)]
 #[test]
-fn test_cat_symlink() {
+fn test_print_symlink() {
     let test_env = TestEnvironment::default();
     test_env.jj_cmd_ok(test_env.env_root(), &["git", "init", "repo"]);
     let repo_path = test_env.env_root().join("repo");
@@ -112,7 +107,7 @@ fn test_cat_symlink() {
     std::os::unix::fs::symlink("symlink1_target", repo_path.join("symlink1")).unwrap();
 
     // Can print multiple files
-    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["cat", "."]);
+    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["file", "print", "."]);
     insta::assert_snapshot!(stdout, @r###"
     c
     a